// src/components/UserProfile/_types/FormHookValues.type.ts
export type FormHookValues = {
  name: string;
  username: string;
  professionalMode: boolean;
  nickname?: string | undefined;
  bio?: string | undefined;
  jopTitle?: string | undefined;
  education?: string | undefined;
  gender?: "" | "MALE" | "FEMALE" | undefined;
  city?: string | undefined;
  facebook?: string | undefined;
  x?: string | undefined;
  linkedIn?: string | undefined;
  instagram?: string | undefined;
  github?: string | undefined;
};